Fix do cast de UInt64[] para UInt32[]
Os metodos dos histogramas estavao a devolver long e o eval so aceitava int

+++ SS_OpenCV/HistogramGray.cs
@@ -13,7 +13,7 @@ namespace SS_OpenCV
 {
     public partial class HistogramGray : Form
     {
-        public HistogramGray(long[] array)
+        public HistogramGray(int[] array)
         {
             InitializeComponent();
